refactor: separate slides logic into its own component, `SlidesCarousel`
